feat: implementing exception handling for deleteUser method

+++ src/main/java/com/devgomes/project/services/UserService.java
@@ -2,13 +2,17 @@ package com.devgomes.project.services;
 
 import com.devgomes.project.entities.User;
 import com.devgomes.project.repositories.UserRepository;
+import com.devgomes.project.services.exceptions.DatabaseException;
 import com.devgomes.project.services.exceptions.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class UserService {
@@ -28,8 +32,16 @@ public class UserService {
         return repository.save(user);
     }
 
-    public void deleteUser(Long id){
-        repository.deleteById(id);
+    public void deleteUser(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            log.error("Integrity violation by deleting user {}: {}", id, e.getMessage());
+            throw new DatabaseException("Can't delete an user that has pending orders.");
+        }
     }
 
     public User updateUser(Long id, User user){
